# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Basket Planner API
  description: |
    API REST para el sistema de planificación de entrenamientos de baloncesto.
    
    ## Autenticación
    La API utiliza autenticación basada en JWT (Bearer Token). Para acceder a los endpoints protegidos,
    incluye el token en el header `Authorization: Bearer <token>`.
    
    ## Roles del sistema
    - **admin**: Acceso completo a todas las funcionalidades
    - **technical_director**: Gestión de su club y equipos
    - **coach**: Gestión de sus equipos y entrenamientos
    - **player**: Acceso a sus planificaciones y feedback
    - **user**: Usuario base sin rol específico
    
    ## Sistema de recomendación
    El sistema utiliza un modelo TensorFlow TFRS (Two-Tower) para recomendar ejercicios
    personalizados según objetivos, intensidad y equipamiento disponible.
  version: 1.0.0
  contact:
    name: Basket Planner Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
  - url: https://api.basket-planner.com/api
    description: Servidor de producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en el login
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              examples: VALIDATION_ERROR
            message:
              type: string
              examples: Validation failed
            details:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                  message:
                    type: string
...
    RecommenderStatus:
      type: object
      properties:
        activeVersion:
          type: string
        modelType:
          type: string
        isLoaded:
          type: boolean
        metrics:
          type: object
          properties:
            accuracy:
              type: number
            coverage:
              type: number
            latencyMs:
              type: number

    # ==================== USER CLUB SCHEMAS ====================
...
  responses:
...
    Unauthorized:
      description: No autenticado - Token JWT inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            error:
              code: UNAUTHORIZED
              message: "Token not provided or invalid"
    Forbidden:
      description: No autorizado - Permisos insuficientes para esta acción
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            error:
              code: FORBIDDEN
              message: "You don't have permission to perform this action"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            error:
              code: NOT_FOUND
              message: "Resource not found"
...
security:
  - BearerAuth: []
paths:
...
  /recommender/status:
    get:
      summary: Estado del recomendador
      description: |
        Obtiene el estado actual del sistema de recomendación.
        
        **Caso de uso:** CU.028 - Sistema recomendador
        
        **Roles permitidos:** admin
      tags: [Recommender]
      responses:
        '200':
          description: Estado del recomendador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommenderStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /recommender/versions:
    get:
      summary: Listar versiones del modelo
      description: |
        Lista las versiones disponibles del modelo de recomendación.
        
        **Roles permitidos:** admin
      tags: [Recommender]
      responses:
        '200':
          description: Lista de versiones
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    createdAt:
                      type: string
                    algorithm:
                      type: string
                    isActive:
                      type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /recommender/train:
    post:
      summary: Entrenar modelo
      description: |
        Inicia un nuevo entrenamiento del modelo de recomendación.
        
        **Roles permitidos:** admin
        
        **⚠️ Operación de larga duración**
      tags: [Recommender]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  type: object
                  description: Configuración de entrenamiento
      responses:
        '202':
          description: Entrenamiento iniciado
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                  status:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /recommender/jobs/{jobId}:
    get:
      summary: Estado del trabajo de entrenamiento
      description: |
        Obtiene el estado de un trabajo de entrenamiento.
        
        **Roles permitidos:** admin
      tags: [Recommender]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estado del trabajo
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, running, completed, failed]
                  progress:
                    type: number
                  error:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== USER CLUBS ENDPOINTS ====================
...